rename zustand -> stores
remove unneeded imports

update deviceinfo hook
