+++ src/components/Global/components/horizontal-list.tsx
@@ -1,6 +1,6 @@
+import useDeviceProfile from '../../../stores/device-profile'
 import { warmItemContext } from '../../../hooks/use-item-context'
 import { useJellifyContext } from '../../../providers'
-import { useDeviceProfileContext } from '../../../providers/Settings'
 import { BaseItemDto } from '@jellyfin/sdk/lib/generated-client/models/base-item-dto'
 import { FlashList, FlashListProps, ViewToken } from '@shopify/flash-list'
 import React, { useRef } from 'react'
@@ -18,7 +18,7 @@ export default function HorizontalCardList({
 }: HorizontalCardListProps): React.JSX.Element {
 	const { api, user } = useJellifyContext()
 
-	const deviceProfile = useDeviceProfileContext()
+	const deviceProfile = useDeviceProfile()
 
 	const onViewableItemsChangedRef = useRef(
 		({ viewableItems }: { viewableItems: ViewToken<BaseItemDto>[] }) => {
