+++ src/components/jellify.tsx
@@ -20,7 +20,7 @@ import Toast from 'react-native-toast-message'
 import JellifyToastConfig from '../constants/toast.config'
 import { useColorScheme } from 'react-native'
 import { CarPlayProvider } from '../providers/CarPlay'
-import { useSelectPlayerEngine } from '../zustand/engineStore'
+import { useSelectPlayerEngine } from '../stores/player-engine'
 /**
  * The main component for the Jellify app. Children are wrapped in the {@link JellifyProvider}
  * @returns The {@link Jellify} component
