--- src/providers/Settings/enums/query-keys.ts
@@ -1,7 +1,0 @@
-const QUERY_KEY_PREFEX = 'SETTINGS'
-
-enum SettingsQueryKey {
-	DeviceProfile = `${QUERY_KEY_PREFEX}_DEVICE_PROFILE`,
-}
-
-export default SettingsQueryKey
